+++ src/main/java/model/service/ReserveService.java
@@ -9,10 +9,10 @@ import java.util.Optional;
 
 public interface ReserveService {
     ReserveDto save(ReserveDto reserve);
-    Optional<ReserveDto> findById(int id);
-    Optional<ReserveDto> update(int id, ReserveDto reserve);
+    Optional<ReserveDto> findById(Long id);
+    Optional<ReserveDto> update(Long id, ReserveDto reserve);
     List<ReserveDto> findAll();
     List<ReserveDto> findByClient(ClientDto client);
     List<ReserveDto> findByDate(LocalDate date);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
